requirements.txt
pandas==2.2.0
beautifulsoup4==4.12.3
requests==2.31.0
polygon-api-client==1.13.3
python-dotenv==1.0.1
flask==3.0.0
flask-cors==4.0.0
langchain==0.1.0
finnhub-python==2.4.19
